requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
dateparser==1.2.0
pymongo==4.7.2
 
ultralytics>=8.0.0,<9.0.0
opencv-python-headless>=4.8.0
av
PySide6>=6.5.0
python-vlc>=3.0.18122
kafka-python
